fix: CommonItemTile
